requirements.txt
azure-keyvault-secrets>=4.10.0
azure-identity>=1.25.0
requests>=2.32.0